% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{create_traffic_policy}
\alias{create_traffic_policy}
\title{Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example}
\usage{
create_traffic_policy(
  Name = "string",
  Document = "string",
  Comment = "string"
)
}
\arguments{
\item{Name}{[required] The name of the traffic policy.}

\item{Document}{[required] The definition of this traffic policy in JSON format. For more information, see \href{http://docs.aws.amazon.com/Route53/latest/APIReference/api-policies-traffic-policy-document-format.html}{Traffic Policy Document Format}.}

\item{Comment}{(Optional) Any comments that you want to include about the traffic policy.}
}
\description{
Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com).
}
\examples{

}
