% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{delete_query_logging_config}
\alias{delete_query_logging_config}
\title{Deletes a configuration for DNS query logging}
\usage{
delete_query_logging_config(Id)
}
\arguments{
\item{Id}{[required] The ID of the configuration that you want to delete.}
}
\description{
Deletes a configuration for DNS query logging. If you delete a configuration, Amazon Route 53 stops sending query logs to CloudWatch Logs. Route 53 doesn't delete any logs that are already in CloudWatch Logs.
}
\details{
For more information about DNS query logs, see CreateQueryLoggingConfig.
}
\section{Accepted Parameters}{
\preformatted{delete_query_logging_config(
  Id = "string"
)
}
}

