% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{delete_reusable_delegation_set}
\alias{delete_reusable_delegation_set}
\title{Deletes a reusable delegation set}
\usage{
delete_reusable_delegation_set(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The ID of the reusable delegation set that you want to delete.}
}
\description{
Deletes a reusable delegation set.
}
\details{
You can delete a reusable delegation set only if it isn't associated with any hosted zones.

To verify that the reusable delegation set is not associated with any hosted zones, submit a GetReusableDelegationSet request and specify the ID of the reusable delegation set that you want to delete.
}
\examples{

}
