% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{delete_traffic_policy}
\alias{delete_traffic_policy}
\title{Deletes a traffic policy}
\usage{
delete_traffic_policy(Id, Version)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy that you want to delete.}

\item{Version}{[required] The version number of the traffic policy that you want to delete.}
}
\description{
Deletes a traffic policy.
}
\section{Accepted Parameters}{
\preformatted{delete_traffic_policy(
  Id = "string",
  Version = 123
)
}
}

