% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_health_check}
\alias{get_health_check}
\title{Gets information about a specified health check}
\usage{
get_health_check(
  HealthCheckId = "string"
)
}
\arguments{
\item{HealthCheckId}{[required] The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.}
}
\description{
Gets information about a specified health check.
}
\examples{

}
