% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_health_check_count}
\alias{get_health_check_count}
\title{Retrieves the number of health checks that are associated with the current AWS account}
\usage{
get_health_check_count()
}
\description{
Retrieves the number of health checks that are associated with the current AWS account.
}
\examples{

}
