% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_hosted_zone_count}
\alias{get_hosted_zone_count}
\title{Retrieves the number of hosted zones that are associated with the current AWS account}
\usage{
get_hosted_zone_count()
}
\description{
Retrieves the number of hosted zones that are associated with the current AWS account.
}
\section{Accepted Parameters}{
\preformatted{get_hosted_zone_count()
}
}

