% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_query_logging_config}
\alias{get_query_logging_config}
\title{Gets information about a specified configuration for DNS query logging}
\usage{
get_query_logging_config(Id)
}
\arguments{
\item{Id}{[required] The ID of the configuration for DNS query logging that you want to get information about.}
}
\description{
Gets information about a specified configuration for DNS query logging.
}
\details{
For more information about DNS query logs, see CreateQueryLoggingConfig and \href{http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/query-logs.html}{Logging DNS Queries}.
}
\section{Accepted Parameters}{
\preformatted{get_query_logging_config(
  Id = "string"
)
}
}

