% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_reusable_delegation_set}
\alias{get_reusable_delegation_set}
\title{Retrieves information about a specified reusable delegation set, including the four name servers that are assigned to the delegation set}
\usage{
get_reusable_delegation_set(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The ID of the reusable delegation set that you want to get a list of name servers for.}
}
\description{
Retrieves information about a specified reusable delegation set, including the four name servers that are assigned to the delegation set.
}
\examples{

}
