% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_traffic_policy}
\alias{get_traffic_policy}
\title{Gets information about a specific traffic policy version}
\usage{
get_traffic_policy(Id, Version)
}
\arguments{
\item{Id}{[required] The ID of the traffic policy that you want to get information about.}

\item{Version}{[required] The version number of the traffic policy that you want to get information about.}
}
\description{
Gets information about a specific traffic policy version.
}
\section{Accepted Parameters}{
\preformatted{get_traffic_policy(
  Id = "string",
  Version = 123
)
}
}

