% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{get_traffic_policy_instance_count}
\alias{get_traffic_policy_instance_count}
\title{Gets the number of traffic policy instances that are associated with the current AWS account}
\usage{
get_traffic_policy_instance_count()
}
\description{
Gets the number of traffic policy instances that are associated with the current AWS account.
}
\examples{

}
