% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{list_geo_locations}
\alias{list_geo_locations}
\title{Retrieves a list of supported geographic locations}
\usage{
list_geo_locations(StartContinentCode = NULL, StartCountryCode = NULL,
  StartSubdivisionCode = NULL, MaxItems = NULL)
}
\arguments{
\item{StartContinentCode}{The code for the continent with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if \code{IsTruncated} is true, and if \code{NextContinentCode} from the previous response has a value, enter that value in \code{startcontinentcode} to return the next page of results.

Include \code{startcontinentcode} only if you want to list continents. Don't include \code{startcontinentcode} when you're listing countries or countries with their subdivisions.}

\item{StartCountryCode}{The code for the country with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if \code{IsTruncated} is \code{true}, and if \code{NextCountryCode} from the previous response has a value, enter that value in \code{startcountrycode} to return the next page of results.

Route 53 uses the two-letter country codes that are specified in \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO standard 3166-1 alpha-2}.}

\item{StartSubdivisionCode}{The code for the subdivision (for example, state or province) with which you want to start listing locations that Amazon Route 53 supports for geolocation. If Route 53 has already returned a page or more of results, if \code{IsTruncated} is \code{true}, and if \code{NextSubdivisionCode} from the previous response has a value, enter that value in \code{startsubdivisioncode} to return the next page of results.

To list subdivisions of a country, you must include both \code{startcountrycode} and \code{startsubdivisioncode}.}

\item{MaxItems}{(Optional) The maximum number of geolocations to be included in the response body for this request. If more than \code{maxitems} geolocations remain to be listed, then the value of the \code{IsTruncated} element in the response is \code{true}.}
}
\description{
Retrieves a list of supported geographic locations.
}
\details{
Countries are listed first, and continents are listed last. If Amazon Route 53 supports subdivisions for a country (for example, states or provinces), the subdivisions for that country are listed in alphabetical order immediately after the corresponding country.
}
\section{Accepted Parameters}{
\preformatted{list_geo_locations(
  StartContinentCode = "string",
  StartCountryCode = "string",
  StartSubdivisionCode = "string",
  MaxItems = "string"
)
}
}

