% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{list_tags_for_resources}
\alias{list_tags_for_resources}
\title{Lists tags for up to 10 health checks or hosted zones}
\usage{
list_tags_for_resources(
  ResourceType = "healthcheck"|"hostedzone",
  ResourceIds = list(
    "string"
  )
)
}
\arguments{
\item{ResourceType}{[required] The type of the resources.
\itemize{
\item The resource type for health checks is \code{healthcheck}.
\item The resource type for hosted zones is \code{hostedzone}.
}}

\item{ResourceIds}{[required] A complex type that contains the ResourceId element for each resource for which you want to get a list of tags.}
}
\description{
Lists tags for up to 10 health checks or hosted zones.
}
\details{
For information about using tags for cost allocation, see \href{http://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html}{Using Cost Allocation Tags} in the \emph{AWS Billing and Cost Management User Guide}.
}
\examples{

}
