% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{list_vpc_association_authorizations}
\alias{list_vpc_association_authorizations}
\title{Gets a list of the VPCs that were created by other accounts and that can be associated with a specified hosted zone because you've submitted one or more \code{CreateVPCAssociationAuthorization} requests}
\usage{
list_vpc_association_authorizations(
  HostedZoneId = "string",
  NextToken = "string",
  MaxResults = "string"
)
}
\arguments{
\item{HostedZoneId}{[required] The ID of the hosted zone for which you want a list of VPCs that can be associated with the hosted zone.}

\item{NextToken}{\emph{Optional}: If a response includes a \code{NextToken} element, there are more VPCs that can be associated with the specified hosted zone. To get the next page of results, submit another request, and include the value of \code{NextToken} from the response in the \code{nexttoken} parameter in another \code{ListVPCAssociationAuthorizations} request.}

\item{MaxResults}{\emph{Optional}: An integer that specifies the maximum number of VPCs that you want Amazon Route 53 to return. If you don't specify a value for \code{MaxResults}, Route 53 returns up to 50 VPCs per page.}
}
\description{
Gets a list of the VPCs that were created by other accounts and that can be associated with a specified hosted zone because you've submitted one or more \code{CreateVPCAssociationAuthorization} requests.
}
\details{
The response includes a \code{VPCs} element with a \code{VPC} child element for each VPC that can be associated with the hosted zone.
}
\examples{

}
