% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_package.R
\docType{package}
\name{paws.route53-package}
\alias{paws.route53}
\alias{paws.route53-package}
\title{paws.route53: Amazon Route 53}
\description{

}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
