% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.route53_operations.R
\name{update_traffic_policy_comment}
\alias{update_traffic_policy_comment}
\title{Updates the comment for a specified traffic policy version}
\usage{
update_traffic_policy_comment(
  Id = "string",
  Version = 123,
  Comment = "string"
)
}
\arguments{
\item{Id}{[required] The value of \code{Id} for the traffic policy that you want to update the comment for.}

\item{Version}{[required] The value of \code{Version} for the traffic policy that you want to update the comment for.}

\item{Comment}{[required] The new comment for the specified traffic policy and version.}
}
\description{
Updates the comment for a specified traffic policy version.
}
\examples{

}
